GET-584 Remove skills builder v2 feature flag
jira-ticket: https://dfedigital.atlassian.net/browse/GET-282

* Remove feature flag for skills builder v2
* Remove old views for skills builder v1
* Remove current job id from user session as that was only used
for v1
* Remove versioning from sessions as that was only used for
skills builder v1/v2
* tidy up all features and specs to only use v2 as the default
