+++ app/models/user_session.rb
@@ -6,8 +6,6 @@ class UserSession
     job_profile_skills
     job_profile_ids
     postcode
-    current_job_id
-    version
   ].freeze
 
   def self.merge_sessions(new_session:, previous_session_data:)
@@ -17,17 +15,9 @@ class UserSession
   def initialize(session)
     @session = session
 
-    # TODO: This should be removed after we go live with skills builder v2
-    @session.clear unless version == expected_version
-
     @session[:visited_pages] ||= []
     @session[:job_profile_skills] ||= {}
     @session[:job_profile_ids] ||= []
-    @session[:version] ||= expected_version
-  end
-
-  def version
-    session[:version]
   end
 
   def postcode
@@ -56,14 +46,6 @@ class UserSession
     session[:registration_triggered_path] = url
   end
 
-  def current_job_id
-    session[:current_job_id]
-  end
-
-  def current_job_id=(value)
-    session[:current_job_id] = value
-  end
-
   def job_profile_skills
     session[:job_profile_skills]
   end
@@ -82,10 +64,6 @@ class UserSession
     session[:visited_pages] << page_key unless page_visited?(page_key)
   end
 
-  def current_job?
-    session[:current_job_id].present?
-  end
-
   def page_visited?(page_key)
     session[:visited_pages].include?(page_key)
   end
@@ -99,24 +77,14 @@ class UserSession
   end
 
   def skill_ids
-    return skill_ids_for_profile(current_job_id) if current_job?
-
     job_profile_skills.values.flatten.uniq
   end
 
   def job_profile_ids
-    return [current_job_id] if current_job?
-
     session[:job_profile_ids]
   end
 
   def skill_ids_for_profile(id)
     job_profile_skills[id.to_s] || []
-  end
-
-  private
-
-  def expected_version
-    Flipflop.skills_builder_v2? ? 2 : 1
   end
 end
