+++ app/views/check_your_skills/index.html.erb
@@ -11,7 +11,7 @@
 
 <div class="govuk-grid-row govuk-!-margin-top-7">
   <div class="govuk-grid-column-two-thirds" >
-    <% if Flipflop.skills_builder_v2? && user_session.job_profile_skills? %>
+    <% if user_session.job_profile_skills? %>
       <%= render "shared/search/form", search_path: check_your_skills_path, scope: 'check_your_skills.index.additional_skills' %>
     <% else %>
       <%= render "shared/search/form", search_path: check_your_skills_path, scope: 'check_your_skills.index' %>
