+++ app/views/skills/_job_profiles_with_skills.html.erb
