--- app/views/skills/v2/index.html.erb
@@ -1,31 +1,0 @@
-<% page_title :skills_index %>
-<% content_for :breadcrumb do
-    generate_breadcrumbs(
-      t('breadcrumb.your_skills'),
-      [
-        [t('breadcrumb.task_list_home'), task_list_path],
-        [t('breadcrumb.check_your_skills'), check_your_skills_path],
-        [t('breadcrumb.search_results'), results_check_your_skills_path(search: params[:search])],
-        [t('breadcrumb.current_job_skills'), job_profile_skills_path(job_profile_id: params[:job_profile_id] || @job_profiles_with_skills.last[:profile_slug], search: params[:search])]
-      ]
-    )
-  end
-%>
-
-<div class="govuk-grid-row govuk-!-margin-top-7 govuk-body">
-  <div class="govuk-grid-column-two-thirds">
-    <div class="govuk-grid-column-full govuk-!-padding-0">
-      <h1 class="govuk-heading-xl">Your skills</h1>
-      <% if @job_profiles_with_skills.any? %>
-        <%= render 'skills/v2/job_profiles_with_skills', job_profiles_with_skills: @job_profiles_with_skills %>
-      <% else %>
-        <p class='govuk-body'>You haven't selected any skills. Tell us a job you've done so we can check what skills you have.</p>
-        <%= link_to('Check my skills', check_your_skills_path, class: 'govuk-link') %>
-      <% end %>
-    </div>
-  </div>
-  <div class="govuk-grid-column-one-third">
-    <%= render 'shared/contact_us' %>
-    <%= render 'shared/save_or_return_to_your_results' %>
-  </div>
-</div>
