+++ spec/features/user_registration_spec.rb
@@ -39,7 +39,7 @@ RSpec.feature 'User registration' do
   end
 
   before do
-    enable_feature!(:user_authentication, :skills_builder_v2)
+    enable_feature!(:user_authentication)
     allow(Notifications::Client).to receive(:new).and_return(client)
   end
