+++ spec/features/user_sign_in_through_registration_spec.rb
@@ -35,7 +35,7 @@ RSpec.feature 'User sign in through registration' do
   end
 
   before do
-    enable_feature!(:user_authentication, :skills_builder_v2)
+    enable_feature!(:user_authentication)
     allow(Notifications::Client).to receive(:new).and_return(client)
   end
