+++ spec/models/user_sessions_spec.rb
@@ -5,11 +5,6 @@ RSpec.describe UserSession do
 
   let(:session) { create_fake_session({}) }
 
-  before do
-    disable_feature! :skills_builder_v2
-    session[:version] = 1
-  end
-
   describe '.merge_sessions' do
     it 'merges old data into new session for selected keys' do
       old_session = {
@@ -24,78 +19,14 @@ RSpec.describe UserSession do
 
     it 'overrides set values in current session with old ones' do
       old_session = {
-        'version' => 5,
         'job_profile_skills' => { '11' => [2, 3, 5] }
       }
       new_session = {
-        'version' => 2,
         'job_profile_skills' => { '11' => [4] }
       }
       described_class.merge_sessions(new_session: new_session, previous_session_data: old_session)
 
-      expect(new_session).to eq('version' => 5, 'job_profile_skills' => { '11' => [2, 3, 5] })
-    end
-  end
-
-  describe '#version' do
-    it 'sets version on new session' do
-      user_session = described_class.new(create_fake_session(session, versioned: false))
-
-      expect(user_session.version).to eq(1)
-    end
-
-    it 'retains correct version on existing session' do
-      session = { version: 1 }
-      user_session = described_class.new(create_fake_session(session, versioned: false))
-
-      expect(user_session.version).to eq(1)
-    end
-
-    it 'retains session when version correct on existing session' do
-      session = {
-        version: 1,
-        job_profile_skills: { '11' => [2, 3, 5] }
-      }
-      user_session = described_class.new(create_fake_session(session, versioned: false))
-
-      expect(user_session.job_profile_skills).to eq('11' => [2, 3, 5])
-    end
-
-    it 'resets version when existing version does not equal current version' do
-      session = { version: 0 }
-      user_session = described_class.new(create_fake_session(session, versioned: false))
-
-      expect(user_session.version).to eq(1)
-    end
-
-    it 'reset session when existing session has wrong version' do
-      session = {
-        version: 0,
-        job_profile_skills: { '11' => [2, 3, 5] }
-      }
-      user_session = described_class.new(create_fake_session(session, versioned: false))
-
-      expect(user_session.job_profile_skills).to be_empty
-    end
-
-    it 'reset session when existing session does not have a version' do
-      session = { job_profile_skills: { '11' => [2, 3, 5] } }
-      user_session = described_class.new(create_fake_session(session, versioned: false))
-
-      expect(user_session.job_profile_skills).to be_empty
-    end
-
-    context 'when skills builder v2 is enabled' do
-      it 'reset session when existing session has wrong version' do
-        enable_feature! :skills_builder_v2
-        session = {
-          version: 1,
-          job_profile_skills: { '11' => [2, 3, 5] }
-        }
-        user_session = described_class.new(create_fake_session(session, versioned: false))
-
-        expect(user_session.job_profile_skills).to be_empty
-      end
+      expect(new_session).to eq('job_profile_skills' => { '11' => [2, 3, 5] })
     end
   end
 
@@ -148,18 +79,6 @@ RSpec.describe UserSession do
       user_session.registration_triggered_path = nil
 
       expect(user_session.registration_triggered_path).to eq('some-path')
-    end
-  end
-
-  describe '#current_job_id' do
-    it 'returns current_job_id value if set' do
-      user_session.current_job_id = 1
-
-      expect(user_session.current_job_id).to eq(1)
-    end
-
-    it 'returns nil if no postcode set' do
-      expect(user_session.current_job_id).to be_nil
     end
   end
 
@@ -233,39 +152,6 @@ RSpec.describe UserSession do
     end
   end
 
-  describe '#current_job?' do
-    context 'when current_job_id key is on the session' do
-      let(:session) {
-        create_fake_session(
-          current_job_id: 12,
-          job_profile_skills: {
-            '11' => [2, 3, 5],
-            '12' => [2, 9, 4]
-          }
-        )
-      }
-
-      it 'returns true when current_job_id key is on the session' do
-        expect(user_session.current_job?).to be true
-      end
-    end
-
-    context 'when current_job_id key is not on the session' do
-      let(:session) {
-        create_fake_session(
-          job_profile_skills: {
-            '11' => [2, 3, 5],
-            '12' => [2, 9, 4]
-          }
-        )
-      }
-
-      it 'returns false when current_job_id key is not on the session' do
-        expect(user_session.current_job?).to be false
-      end
-    end
-  end
-
   describe '#page_visited?' do
     it 'returns true if the page has alreaady been persisted on the session' do
       user_session.track_page('some_page')
@@ -309,76 +195,40 @@ RSpec.describe UserSession do
   end
 
   describe '#skill_ids' do
-    context 'when there is a current_job_id on the session and we are using Skills Builder v1' do
-      let(:session) {
-        create_fake_session(
-          current_job_id: 11,
-          job_profile_skills: {
-            '11' => [2, 3, 5],
-            '12' => [2, 9, 4]
-          }
-        )
-      }
+    let(:session) {
+      create_fake_session(
+        job_profile_skills: {
+          '11' => [2, 3, 5],
+          '12' => [2, 9, 4]
+        }
+      )
+    }
 
-      it 'returns the skill ids on the session that belong to job profile id: 11' do
-        expect(user_session.skill_ids).to contain_exactly(2, 3, 5)
-      end
-    end
-
-    context 'when there is no current_job_id on the session' do
-      let(:session) {
-        create_fake_session(
-          job_profile_skills: {
-            '11' => [2, 3, 5],
-            '12' => [2, 9, 4]
-          }
-        )
-      }
-
-      it 'returns the all skill ids on the session' do
-        expect(user_session.skill_ids).to contain_exactly(2, 3, 5, 9, 4)
-      end
+    it 'returns the all skill ids on the session' do
+      expect(user_session.skill_ids).to contain_exactly(2, 3, 5, 9, 4)
     end
   end
 
   describe '#job_profile_ids' do
-    context 'when there is a current_job_id on the session and we are using Skills Builder v1' do
-      let(:session) {
-        create_fake_session(
-          current_job_id: 12,
-          job_profile_skills: {
-            '11' => [2, 3, 5],
-            '12' => [2, 9, 4]
-          }
-        )
-      }
+    let(:session) {
+      create_fake_session(
+        job_profile_ids: [11, 12],
+        job_profile_skills: {
+          '11' => [2, 3, 5],
+          '12' => [2, 9, 4]
+        }
+      )
+    }
 
-      it 'returns just the current job profile id on the session' do
-        expect(user_session.job_profile_ids).to contain_exactly(12)
-      end
+    it 'returns all the job profile ids on the session' do
+      expect(user_session.job_profile_ids).to contain_exactly(11, 12)
     end
 
-    context 'when there is no current_job_id on the session' do
-      let(:session) {
-        create_fake_session(
-          job_profile_ids: [11, 12],
-          job_profile_skills: {
-            '11' => [2, 3, 5],
-            '12' => [2, 9, 4]
-          }
-        )
-      }
-
-      it 'returns all the job profile ids on the session' do
-        expect(user_session.job_profile_ids).to contain_exactly(11, 12)
-      end
-
-      it 'does not add duplicates do job profile ids' do
-        user_session.set_skills_ids_for_profile(11, [2, 3])
-        user_session.set_skills_ids_for_profile(11, [2, 5])
-        user_session.set_skills_ids_for_profile(11, [2, 3, 5])
-        expect(user_session.job_profile_ids).to contain_exactly(11, 12)
-      end
+    it 'does not add duplicates do job profile ids' do
+      user_session.set_skills_ids_for_profile(11, [2, 3])
+      user_session.set_skills_ids_for_profile(11, [2, 5])
+      user_session.set_skills_ids_for_profile(11, [2, 3, 5])
+      expect(user_session.job_profile_ids).to contain_exactly(11, 12)
     end
   end
